// src/app/models/respuesta.qr.model.ts
export class RespuestaQr{

    existe_usuario: string;
    titulo_1: string;
    id: number;
    nombre: string;
    descripcion: string;
    fecha_creacion: Date
    codigo_qr: string;
    url_qr: string;
    serial: string;
    razon_social_id: number;
    usuario_id: number;
    tipo_estado_id: number;
    tipo_producto_id: number;
    fecha_baja: null;
    urlimg: string;
    condicion: number;
    titulo_2: string;
    direccion: string;
    telefono: string;
    correo: string;
    titulo_3: string;
    email: string;
    rol: string;
    fecha_alta: Date
    estado: number;
    genero: string;
    telcel:string;
    telref:string;
    idempresa:number;
    nombre_rz:string;


}